requirements.txt
asgiref==3.2.10
certifi==2020.4.5.1
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.2
django-cors-headers==3.4.0
django-heroku==0.3.1
django-markwhat==1.6.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
docopt==0.6.2
gunicorn==20.0.4
idna==2.9
psycopg2==2.8.5
python-dotenv==0.13.0
pytz==2020.1
six==1.14.0
sqlparse==0.4.1
static3==0.7.0
urllib3==1.25.9
whitenoise==5.2.0
